+++ src/ui/src/js/services/garden_service.js
@@ -9,12 +9,56 @@ gardenService.$inject = ['$rootScope', '$http'];
 export default function gardenService($rootScope, $http) {
   const GardenService = {};
 
+  const minimum_fields = [
+                "name",
+                "has_parent",
+                "parent",
+                "version",
+                "has_parent",
+                "shared_users",
+                "default_user",
+                "connection_type",
+                "metadata",
+                // Have to list all fields since instances is called out
+                "systems__id",
+                "systems__name",
+                "systems__description",
+                "systems__version",
+                "systems__icon_name",
+                "systems__display_name",
+                "systems__metadata",
+                "systems__namespace",
+                "systems__local",
+                "systems__template",
+                "systems__groups",
+                "systems__prefix_topic",
+                "systems__requires",
+                "systems__requires_timeout",
+                "systems__commands",
+                "systems__instances__id",
+                "systems__instances__name",
+                // Skipping instance status info
+                "systems__instances__status",
+                "systems__instances__description",
+                "systems__instances__queue_type",
+                "systems__instances__queue_info",
+                "systems__instances__icon_name",
+                "systems__instances__metadata",
+                "receiving_connections__api",
+                "receiving_connections__status",
+                "publishing_connections__api",
+                "publishing_connections__status",];
+
   GardenService.getGardens = function() {
     return $http.get('api/v1/gardens/');
   };
 
   GardenService.getGarden = function(name) {
-    return $http.get('api/v1/gardens/' + encodeURIComponent(name));
+    return $http.get('api/v1/gardens/' + encodeURIComponent(name), {
+      params: {
+        include_fields: minimum_fields.join()
+      },
+    });
   };
 
   GardenService.syncGardens = function() {