+++ src/app/beer_garden/requests.py
@@ -1261,7 +1261,6 @@ def handle_event(event):
                     "status_updated_at",
                     "target_garden",
                     "updated_at",
-                    "command_type",
                     "metadata",
                 ):
                     new_value = getattr(event.payload, field)