+++ CHANGELOG.md
@@ -9,7 +9,7 @@ TBD
 - Updated Instance Status icon for Downstream Instances to reflect error states (#1907)
 - Updated UI Lazy Loading (#1907)
 - Fixed bug in local garden loading all gardens to one hop away (#1905)
-- Expanded Garden query to support include and exclude field parameters
+- Expanded Garden query to support include and exclude field parameters (#1916)
 
 ## 3.30.0rc1