+++ CHANGELOG.md
@@ -9,6 +9,7 @@ TBD
 - Updated Instance Status icon for Downstream Instances to reflect error states (#1907)
 - Updated UI Lazy Loading (#1907)
 - Fixed bug in local garden loading all gardens to one hop away (#1905)
+- Updated request pre and post save methods to set expiration for TEMP and children (#1904)
 
 ## 3.30.0rc1