+++ CHANGELOG.md
@@ -10,6 +10,7 @@ TBD
 - Updated UI Lazy Loading (#1907)
 - Fixed bug in local garden loading all gardens to one hop away (#1905)
 - Updated request pre and post save methods to set expiration for TEMP and children (#1904)
+- Fixed bug where Garden Entry Points went unresponsive (#1908)
 
 ## 3.30.0rc1