+++ CHANGELOG.md
@@ -14,7 +14,7 @@ TBD
 - Updated Request Pre_Save function to check for Orphan by parent ID (#1913)
 - Updated request pre and post save methods to set expiration for TEMP and children (#1904)
 - Fixed bug where Garden Entry Points went unresponsive (#1908)
-- Update UI Events to utilize Instance and System events as partial Garden updates (#)
+- Update UI Events to utilize Instance and System events as partial Garden updates (#1925)
 
 ## 3.30.0rc1