+++ CHANGELOG.md
@@ -15,6 +15,7 @@ TBD
 - Updated request pre and post save methods to set expiration for TEMP and children (#1904)
 - Fixed bug where Garden Entry Points went unresponsive (#1908)
 - Update UI Events to utilize Instance and System events as partial Garden updates (#1925)
+- Removed locking from internal queue management for event handlers (#1921)
 
 ## 3.30.0rc1