+++ CHANGELOG.md
@@ -16,6 +16,7 @@ TBD
 - Fixed bug where Garden Entry Points went unresponsive (#1908)
 - Removed duplicate Garden Sync Event sent to UI from downstream events (#1922)
 - Update UI Events to utilize Instance and System events as partial Garden updates (#1925)
+- Bypass Mongoengine for bulk create/update to improve Request throughput in Mongo (#1920)
 - Removed locking from internal queue management for event handlers (#1921)
 
 ## 3.30.0rc1